+++ lib/packages/core/src/api/factories/message/index.ts
@@ -2,3 +2,5 @@
 
 export * from './sendEncryptedTextMessage';
 export * from './sendTextMessage';
+export * from './sendMessage';
+export * from './sendEncryptedMessage';