+++ lib/packages/core/CHANGELOG.md
@@ -3,10 +3,16 @@ All notable changes to this project will be documented in this file.
 
 ## 0.4.1 (TBR)
 - Initiated work on assets Api
-    - added `getAsset()`
-    - added `getAllAssets()`
+    - added `getAsset`
+    - added `getAllAssets`
 - Fixed encoding issue on BRS requests (#826)
 - Added `getAllContractIds` to contract Api
+- deprecated `sendTextMessage`
+    - will be removed in next version (0.5)
+    - use `sendMessage`
+- deprecated `sendEncryptedTextMessage`
+    - will be removed in next version (0.5)
+    - use `sendEncryptedMessage`
     
 ## 0.4.0
 - Changed License: From GPL-3.0 to Apache 2.0